client: move App component
