+++ website/client/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import App from './App';
+import App from './components/App';
 
 new Vue({ // eslint-disable-line no-new
   el: '#app',
